+++ src/Database/Migrations/2019_03_09_120411_add_dz_user_restricted_ips_table.php
@@ -18,13 +18,13 @@ class AddDzUserRestrictedIpsTable extends Migration {
             $table->bigIncrements('id');
 
             $table->string('source_type')->nullable();
-            $table->unsignedInteger('source_id')->nullable();
+            $table->unsignedBigInteger('source_id')->nullable();
 
             $table->string('ip_address')->nullable();
             $table->string('comments');
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
